+++ amr_viz/src/app/App.tsx
@@ -1,0 +1,5 @@
+import { VizDashboardPage } from "../features/viz/pages/VizDashboardPage";
+
+export default function App() {
+  return <VizDashboardPage />;
+}